+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { Header } from "../containers/header";
+import { Header } from "../containers/Header";
 
 const inter = Inter({ subsets: ["latin"] });
